Extract new element creation into its own factory function
